Adds capability to cache Braze messages
The cache implementation used by BrazeMessages is configurable. This PR
uses an in memory cache (InMemoryCache) so no data will persist across
pageviews. We need to ensure that we're clearing the cache on logout and
consent change on DCR and frontend before switching to the real
localStorage backed cache (LocalMessageCache). This PR includes clearing
data on DCR, but we still need to make the change for frontend.

The delivery behaviour of Braze messages to the browser doesn't exactly
fit our use cases. Once Braze has given us a message via the
subscribeToInAppMessage callback they consider it delivered, and it
won't be delivered again. But there are many reasons why we might not be
able to show a message in that same pageview. For example: banner
timeout limit is reached, a higher priority message is shown instead,
the page isn't suitable (e.g. paid content) or doesn't have the required
slot (e.g. an epic message is delivered on a page without that slot).