Adding a button for a documentcloud.org embedded document that links out to document cloud
Android seems to cause problems for embedded documents in that the gesture based
scrolling and zooming doesnt work well in iframes on that platform.

Scribd works around this problem by having a button that links to the document on
scribd.com where the scrolling and zooming works.

We are moving to documentcloud.org because of privacy issues around scribd
tracking users.

In order to achieve parity with scribd this change adds a button that links
out to documentcloud.

This is not the whole story as on documentcloud.org scrolling works but
zooming doesnt. DocumentCloud are 'working' on this.

To get round this problem we convert the url to the pdf viewer to the
pdf its-self, meaning the pdf gets rendered using the native rendering on
the browser which seems to work fine.